implement test_object
